Refactoring: cleaning the code
* remove unnecessary directives.

* The IdentityDbContext.cs and AppUser.cs files
have been moved to the Infrastructure folder.Identity\Data

* The Identity component services have been moved to the
Infrastructure folder.Identity\Services

* the extension method responsible for connecting Identity
to the WebAPI project has been removed from the
DependencyInjection class of the Infrastructure project
