--- serverapp/Infrastructure.Identity/Interfaces/IUserManager.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Infrastructure.Identity.Interfaces
-{
-    public interface IUserManager
-    {
-        Task SendConfirmationEmail(string userEmeil);
-    }
-}
