+++ serverapp/Infrastructure/Infrastructure.csproj
@@ -2,7 +2,6 @@
 
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
-    <UserSecretsId>493c2e4f-71c7-4731-a543-4e6923ce4b95</UserSecretsId>
   </PropertyGroup>
 
   <ItemGroup>
@@ -24,11 +23,6 @@
     <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="5.0.0" />
     <PackageReference Include="MimeKit" Version="2.11.0" />
     <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="5.0.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Identity\Migrations\" />
-    <Folder Include="Identity\Queries\" />
   </ItemGroup>
 
   <ItemGroup>
