+++ serverapp/WebApi/Startup.cs
@@ -1,21 +1,14 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.HttpsPolicy;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Reflection;
-
 using MediatR;
 
 using Infrastructure;
+using Infrastructure.Identity.DependencyInjection;
 using Persistence;
 using WebApi.Middleware;
 
@@ -47,7 +40,7 @@ namespace WebApi
 
             services.AddControllers();
 
-            services.AddMediatR(typeof(Infrastructure.DependencyInjection).Assembly);
+            services.AddMediatR(typeof(Infrastructure.Identity.DependencyInjection.DependencyInjection).Assembly);
         }
 
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
